refact: tidy up SendNewJobsNotification listener
To simplify this listener, we're now passing a Collection of jobs into it.
This way, we can now use Collection methods such as ->count() and isEmpty().

We're using the new Notification::new() static method to instantiate
our notification client.

The name of the event when a notification is clicked now has it's
prefixed defined as a constant for consistency.