test: create test to assert we're notifying of relevant jobs
These tests ensure that when the JobsPosted event is constructed,
it only contains the jobs which are relevant based on the user's
filter preferences.

Unfortunately we can't mock the NativePHP Notification facade, so
instead we're just ensuring that the jobs within the event contain
what we're expecting to notify the user about.